data_enrichment/requirements.txt
geojson==3.1.0
shapely==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.0
